run_bot.py
"""
DHASU-RUSHER Bot Launcher
Runs Discord bot and multiple Free Fire bots from accounts.txt
WITH DYNAMIC RELOAD - Add/Remove accounts without restart!
"""

import asyncio
import json
import os
import sys
from datetime import datetime
from pathlib import Path
import time

# Environment variables for deployment
DISCORD_TOKEN = os.getenv('DISCORD_BOT_TOKEN')
DISCORD_CHANNEL = os.getenv('DISCORD_CHANNEL_ID')
ACCOUNTS_ENV = os.getenv('ACCOUNTS', '')

# Import both bots
import discord_bot
from main import FreeFireBot

# Import keep-alive for deployment
try:
    from keep_alive import start_health_server
    KEEP_ALIVE_ENABLED = True
except ImportError:
    KEEP_ALIVE_ENABLED = False

# Global tracking
all_bots = {}  # {uid: bot_instance}
bot_tasks = {}  # {uid: task}
accounts_file_path = None
last_modified_time = 0

def load_config():
    """Load configuration from environment variables or config.json"""
    # Try environment variables first (for deployment)
    if DISCORD_TOKEN and DISCORD_CHANNEL:
        print("📡 Loading config from environment variables...")
        return {
            'discord': {
                'bot_token': DISCORD_TOKEN,
                'command_channel_id': DISCORD_CHANNEL,
                'prefix': '/'
            },
            'settings': {
                'auto_reconnect': True,
                'log_discord_commands': True,
                'accounts_file': 'accounts.txt'
            }
        }
    
    # Fallback to config.json (for local development)
    config_path = os.path.join(os.path.dirname(__file__), 'config.json')
    if os.path.exists(config_path):
        print("📁 Loading config from config.json...")
        with open(config_path, 'r') as f:
            return json.load(f)
    
    print("❌ Error: No configuration found!")
    print("   Set environment variables OR create config.json")
    return None

def load_accounts():
    """Load accounts from environment variable or accounts.txt"""
    global accounts_file_path
    
    # Try environment variable first (for deployment)
    if ACCOUNTS_ENV:
        print("📡 Loading accounts from environment variable...")
        accounts = {}
        for account_str in ACCOUNTS_ENV.split(','):
            account_str = account_str.strip()
            if ':' in account_str:
                uid, password = account_str.split(':', 1)
                uid = uid.strip()
                accounts[uid] = {'uid': uid, 'password': password.strip()}
        return accounts
    
    # Fallback to accounts.txt (for local development)
    accounts_file_path = os.path.join(os.path.dirname(__file__), 'accounts.txt')
    if os.path.exists(accounts_file_path):
        print("📁 Loading accounts from accounts.txt...")
        accounts = {}
        with open(accounts_file_path, 'r') as f:
            for line in f:
                line = line.strip()
                if line and ':' in line:
                    uid, password = line.split(':', 1)
                    uid = uid.strip()
                    accounts[uid] = {'uid': uid, 'password': password.strip()}
        return accounts
    
    print("❌ Error: No accounts found!")
    print("   Set ACCOUNTS environment variable OR create accounts.txt")
    return {}

def get_file_modified_time():
    """Get last modified time of accounts.txt"""
    global accounts_file_path
    if accounts_file_path and os.path.exists(accounts_file_path):
        return os.path.getmtime(accounts_file_path)
    return 0

async def start_bot(uid, password):
    """Start a single Free Fire bot"""
    print(f"   🟢 Starting bot for UID: {uid}")
    bot = FreeFireBot(uid=uid, password=password)
    all_bots[uid] = bot
    
    # Create task for this bot
    task = asyncio.create_task(
        bot.run_account(),
        name=f"FF Bot {uid}"
    )
    bot_tasks[uid] = task
    return task

async def stop_bot(uid):
    """Stop a single Free Fire bot"""
    if uid in all_bots:
        print(f"   🔴 Stopping bot for UID: {uid}")
        bot = all_bots[uid]
        await bot.stop()
        
        if uid in bot_tasks:
            bot_tasks[uid].cancel()
            del bot_tasks[uid]
        
        del all_bots[uid]

async def account_file_watcher():
    """Watch accounts.txt for changes and reload bots dynamically"""
    global last_modified_time
    
    print("👁️  File watcher started - monitoring accounts.txt for changes...")
    
    while True:
        try:
            await asyncio.sleep(2)  # Check every 2 seconds
            
            current_modified_time = get_file_modified_time()
            
            # File changed?
            if current_modified_time > last_modified_time and last_modified_time > 0:
                print("\n🔄 accounts.txt changed! Reloading accounts...")
                
                # Load new accounts
                new_accounts = load_accounts()
                current_uids = set(all_bots.keys())
                new_uids = set(new_accounts.keys())
                
                # Find accounts to add
                to_add = new_uids - current_uids
                # Find accounts to remove
                to_remove = current_uids - new_uids
                
                # Remove old accounts
                for uid in to_remove:
                    await stop_bot(uid)
                
                # Add new accounts
                for uid in to_add:
                    account = new_accounts[uid]
                    await start_bot(account['uid'], account['password'])
                
                if to_add or to_remove:
                    print(f"✅ Reload complete! Added: {len(to_add)}, Removed: {len(to_remove)}")
                    print(f"📊 Total active bots: {len(all_bots)}")
                else:
                    print("ℹ️  No changes detected in accounts")
            
            last_modified_time = current_modified_time
            
        except Exception as e:
            print(f"⚠️  File watcher error: {e}")
            await asyncio.sleep(5)

def print_banner():
    """Print startup banner"""
    print("=" * 60)
    print("🤖 DHASU-RUSHER BOT LAUNCHER (Dynamic Reload)")
    print("=" * 60)
    print(f"⏰ Started at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    print("=" * 60)

async def main():
    """Main launcher function"""
    global last_modified_time
    
    print_banner()
    
    # Start health check server for deployment (keeps Render alive)
    if KEEP_ALIVE_ENABLED:
        port = int(os.getenv('PORT', 8080))
        start_health_server(port)
    
    # Load configuration
    config = load_config()
    if not config:
        return
    
    # Load accounts
    accounts = load_accounts()
    if not accounts:
        print("❌ Error: No accounts found in accounts.txt")
        return
    
    # Set initial file modified time
    last_modified_time = get_file_modified_time()
    
    print(f"\n📋 Configuration loaded successfully!")
    print(f"   Discord Channel ID: {config['discord']['command_channel_id']}")
    print(f"   Total Accounts: {len(accounts)}")
    print()
    
    # Validate configuration
    if config['discord']['bot_token'] == "YOUR_DISCORD_BOT_TOKEN_HERE":
        print("❌ Error: Please set your Discord bot token in config.json")
        print("   Edit config.json and add your bot token")
        return
    
    if config['discord']['command_channel_id'] == "YOUR_CHANNEL_ID_HERE":
        print("❌ Error: Please set your Discord channel ID in config.json")
        print("   Edit config.json and add your channel ID")
        return
    
    # Create bot instances for all accounts
    print("🚀 Starting bots...\n")
    
    ff_tasks = []
    for idx, (uid, account) in enumerate(accounts.items(), 1):
        print(f"   [{idx}] Creating bot for UID: {uid}")
        task = await start_bot(account['uid'], account['password'])
        ff_tasks.append(task)
    
    # Create Discord bot task
    discord_task = asyncio.create_task(
        discord_bot.run_discord_bot(config),
        name="Discord Bot"
    )
    
    # Create file watcher task
    watcher_task = asyncio.create_task(
        account_file_watcher(),
        name="File Watcher"
    )
    
    print(f"\n✅ All bots are starting!")
    print("━" * 60)
    print("📢 Commands available in Discord:")
    print("   /lw [teamcode]  - Start level-up")
    print("   /stop [teamcode] - Stop level-up")
    print("   /status - Check bot status")
    print("   /bothelp - Show help")
    print("━" * 60)
    print("🔄 Dynamic Reload ENABLED!")
    print("   → Add account to accounts.txt = Auto start")
    print("   → Remove account = Auto stop")
    print("   → No restart needed!")
    print("━" * 60)
    print("\n⚠️  Press Ctrl+C to stop all bots\n")
    
    try:
        # Run Discord bot, file watcher, and all FF bots concurrently
        all_tasks = [discord_task, watcher_task] + list(bot_tasks.values())
        await asyncio.gather(*all_tasks)
    except KeyboardInterrupt:
        print("\n\n🛑 Shutdown signal received...")
        print("   Stopping Discord bot...")
        discord_task.cancel()
        
        print("   Stopping file watcher...")
        watcher_task.cancel()
        
        print(f"   Stopping {len(all_bots)} Free Fire bot(s)...")
        for uid in list(all_bots.keys()):
            await stop_bot(uid)
        
        print("✅ All bots stopped successfully!")
    except Exception as e:
        print(f"\n❌ Error: {e}")
        discord_task.cancel()
        watcher_task.cancel()
        for uid in list(all_bots.keys()):
            await stop_bot(uid)

if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("\n👋 Goodbye!")


