requirements.txt
requests
httpx
google
protobuf
pycryptodome
psutil
PyJWT
urllib3
protobuf-decoder
pytz
aiohttp
cfonts
flask
discord.py